Fix system freeze when user exits emulator via home menu button
This bug occurred since version 2.0, most likely due to the update of libctru to version 2.x
It's more like a quick fix. There are plenty of aptMainLoop() calls in this app which i haven't noticed in other homebrew apps.
So there is probably a better approach (and a lot more refactoring) to handle this issue.